Search Bar and Color status slide button added.
Added a search bar where you can search the categories (with the posts in it), if you type a certain letter, all categories containing this letter will be shown. Also added the color status slide button, only the admin can switch between the status "color" or "black & grey" for every post. The other users can see whether the status of the post is "color" or "black & grey", but can't edit the status.